docs: fix placeholder docstrings, param-type mismatches, and README gaps
Replaces TODO docstrings on the new segmentation config dataclasses with
real descriptions, corrects a docstring/signature type mismatch and a
missing return type annotation in the stitching module, and clarifies
why Tesseract OCR is a required install.

+++ src/stitching/utils.py
@@ -14,7 +14,7 @@ def _resize_cores(
 
     Args:
         cores (list[Image.Image]): Raw core crops to resize.
-        core_height_px (float): Pixel budget for a core spanning core_height_m metres.
+        core_height_px (int): Pixel budget for a core spanning core_height_m metres.
         core_height_m (float): Depth extent, in metres, that core_height_px pixels represents.
         core_width_rerror (float): Maximum allowed width ratio vs. the reference core before
             a core is treated as an outlier and rescaled using the reference core's scale factor.
